requirements.txt
requests>=2.25.1
pdfplumber>=0.7.0
pdf2image>=1.16.3
Pillow>=10.0.0